Arduino Based Counter
This is an Arduino based embedded system, where we aim to count the 
a number of people present in a certain room at any instance.
 
In this COVID era, social distancing is a very important protocol to be 
followed amongst others. To ensure such physical distancing, it is necessary to 
set a strict limit to the number of people present in a closed room. Our project 
aims to serve this purpose. 
Even though the project has been curated for the COVID situation, it still has 
its utilities in the post-COVID era. This project shall eventually help in 
preventing excessive gathering at any closed place at any time. 
Banks, post offices, shopping malls are some of the very examples of places 
where we can effectively apply our project.